Show recent polls when no active polls
